+++ src/internal/HRTFKernel.h
@@ -24,7 +24,7 @@ class HRTFKernel
 public:
     // Note: this is destructive on the passed in AudioChannel.
     // The length of channel must be a power of two.
-    HRTFKernel(AudioChannel *, size_t fftSize, float sampleRate);
+    HRTFKernel(AudioChannel *, int fftSize, float sampleRate);
 
     HRTFKernel(std::unique_ptr<FFTFrame> fftFrame, float frameDelay, float sampleRate)
         : m_fftFrame(std::move(fftFrame))
@@ -35,7 +35,7 @@ public:
 
     FFTFrame * fftFrame() { return m_fftFrame.get(); }
 
-    size_t fftSize() const { return m_fftFrame->fftSize(); }
+    int fftSize() const { return m_fftFrame->fftSize(); }
     float frameDelay() const { return m_frameDelay; }
 
     // Converts back into impulse-response form.
